fix(YSP-415): no more spinning text
It turns out that GinLB does this animation on purpose it seems.  So we
target the throbber (spinner) and disable the animation.  We also set
the display to inline so that the text displays in a more desirable
location; otherwise the text can be clipped by the other div elements.