fix(YSP-415): messages are clearly visible
While removing the spinning text, it causes the text that did appear for
the user to be obfuscated by the delete and edit links that usually
appear above the image.  This ensures that the throbber appears in front
of those, hiding the buttons and allowing the text to be fully shown.