fix(twig): update default aria label in view block
Set the default aria label to "Unknown view" instead of using the
parentNode value. This ensures a consistent fallback label when
parentNode is not provided.

This is a temporary measure since the parentNode is not the view's
content type.  And looking at kint, there is no reference to that in
what's being passed to here.  More work will be needed to implement
this.