fix(YSP-1052): improve table contrast in dark mode
Removes redundant dark mode table contrast styles from admin-theme.css and
updates layout-builder.css to apply correct background and text color for
odd table rows in dark mode. This ensures better readability and consistent
appearance in the Gin admin theme.

Pairing this with the component library twig work to scope the table
handling as it's taking over too much of the system.  By doing this, my
hope is that we can actually remove some of the existing markup later
that was used to fix this bleeding of CSS.